+++ pyomo/contrib/mindtpy/mip_solve.py
@@ -16,7 +16,8 @@ from pyomo.core import Constraint, Expression, Objective, minimize, value
 from pyomo.opt import TerminationCondition as tc
 from pyomo.opt import SolutionStatus, SolverFactory
 from pyomo.contrib.gdpopt.util import copy_var_list_values, SuppressInfeasibleWarning, _DoNothing, get_main_elapsed_time, time_code
-from pyomo.contrib.gdpopt.mip_solve import distinguish_mip_infeasible_or_unbounded
+from pyomo.contrib.gdpopt.solve_master_problem import (
+    distinguish_mip_infeasible_or_unbounded)
 from pyomo.solvers.plugins.solvers.persistent_solver import PersistentSolver
 from pyomo.common.dependencies import attempt_import
 from pyomo.contrib.mindtpy.util import generate_norm1_objective_function, generate_norm2sq_objective_function, generate_norm_inf_objective_function, generate_lag_objective_function, set_solver_options, GurobiPersistent4MindtPy, update_dual_bound, update_suboptimal_dual_bound
